Fix Spring bean cycle via circuit breaker disable path
ExternalApiCircuitBreaker no longer depends on ExternalDataLayerConfigService
(which pulls LayerProviderRegistry → providers → HTTP clients). Disable
logic moved to ExternalDataLayerAutoDisableService (AppSettings only).
